Add "MdAdmonition" block
Block type implements the "Admonition" extension for python markdown. As this extension is not supported in most markdown implementation, type was defined in a separate "Extensions" namespace to make it more clear that is not a CommonMark feature. The internal implementation is not separated thought as this would have required a large refactoring.