feat: Add generic DocumentSet
Add DocumentSet<T> class which implements the same functionality as MdDocumentSet but
is not tied to MdDocument. "Documents" need to only implement the new (quite small)
IDocument interface. Markdown-specific funtionality of MdDocumentSet was reimplemented
as extension methods for DocumentSet<MdDocument>
Implementation of IDocument was added to MdDocument.
MdDocumentSet still exists and with unchanged implementation but is now marked obsolete.