feat: Allow customization of escaping
Adds ITextFormatter abstraction for escaping text. The formatter that
is used for escaping text during serialization of spans can be
customized by providing a implementation of ITextFormatter and passing
it to serialization via MdSerializationOptions.

The formatter is specified using the newly introduced property
MdSerializationOptions.TextFormatter.

If no formatter is specified, DefaultTextFormatter is used which
escapes exactly like the previous implementation in MdTextSpan.