fix: Prevent modifications of the default serialization options instance
MdSerializationOptions.Default provides access to the default
serialization options. While the 'Default' field is readonly, the
individual settings of the instance could be modified because
MdSerializationOptions provides setters for all properties.

To prevent such modifications without introducing additional
abstractions the property setters will now check if it is allowed to
modify the instance and throw when a trying to set a property of the
default instance.