feat: Add option to increase the indentation in multi-level lists
Add property MdSerializationOptions.MinimumListIndentationWidth to
control indentation in multi-level lists.
List items are always indented at by at least the length of the list
marker (2 for bullet list, 3 or more for ordered lists).
With this change, the indentation is now the maximum of the
list marked length and the specified indentation width.

Default value of the property is 2, so lists serialized using the
default options are indented by the same amount as prior to the
introduction of the option.