build: Fix target that invokes inheritdoc
When building in Azure DevOps, the output directory is not
located wihtin the source directory. Thus, starting 'dotnet tool run'
in the output directory will not use the tool manifest.

To fix this, start the tool in the project directory and pass
in the output directory as base directory for 'inheritdoc' instead