feat: Add methods to copy MdSerializationOptions instances
Add Clone() method to MdSerializationOptions that creates a copy of an otpions instance.
Cloning default instances (which are read-only) removes the read-only flag and returns a instance which's properties can be modified.

In addition to Clone() add the With() method to that allows specifiying a update action (as Action<MdSerializationOptions>) to apply to the copy.